Improve prop definitions
Orientation now only supports 'up' and 'down', any other input will trigger an error in console and alert user of what the supported values are. Same goes for direction.

Adds a custom prop validation to 'percentage', which validate the type and checks if it is between 0 and 100 and makes it required.